Factored a lot of code into three-panel.component.ts
