+++ src/app/pages/details-texts/details-texts.component.ts
@@ -1,8 +1,7 @@
-import { Component, OnInit, SecurityContext, ViewEncapsulation } from '@angular/core';
-import { GetDataService } from '../../services/get-data/get-data.service';
-import { DomSanitizer } from '@angular/platform-browser';
-import { HandleBreadcrumbsService } from '../../services/handle-breadcrumbs/handle-breadcrumbs.service';
-import { Router, ActivatedRoute } from '@angular/router';
+import { Component, ViewEncapsulation } from '@angular/core';
+
+import { Observable } from 'rxjs';
+
 import {
   composedPath,
   splitOutTranslations,
@@ -11,8 +10,8 @@ import {
   findAttributeBy,
   findAncestorByTag
 } from '../../../utils/utils';
-import { DIRECTION, PANEL_TYPE, SESSION_KEYS } from '../../../utils/consts';
 import { environment } from 'src/environments/environment';
+import { ThreePanel } from 'src/app/components/three-panel.component';
 
 @Component({
   selector: 'app-details-texts',
@@ -20,159 +19,46 @@ import { environment } from 'src/environments/environment';
   styleUrls: ['../details/details.component.scss'],
   encapsulation: ViewEncapsulation.None
 })
-export class DetailsTextsComponent implements OnInit {
-  public metadataPanel: any = "<i class='fas fa-spinner'></i>";
-  public chosenTermText: string = "";
-  public currentPage = 1;
-  public textPanel: any;
-  public isMetadataPanelActive = window.innerWidth > 991 ? true : false;
-  public totalPages: any;
-  public paginatedPages: any;
-  public middlePanel: any = "<i class='fas fa-spinner'></i>";
-  public isTextPanelActive = window.innerWidth > 991 ? true : false;
-  public shouldShowPaginationArrows: boolean;
-  private paginationSliceStart: number = 0;
-  private paginationSliceEnd: number = 7;
-  public totalLines: number;
-  private project: string = 'neo';
-  private isMobile: boolean;
+export class DetailsTextsComponent extends ThreePanel {
   private item: string = '';
 
-  constructor(
-    private getDataService: GetDataService,
-    private breadcrumbsService: HandleBreadcrumbsService,
-    private sanitizer: DomSanitizer,
-    private router: Router,
-    private route: ActivatedRoute
-  ) {
-    this.breadcrumbsService.setBreadcrumbs(this.router);
-  }
-
-  ngOnInit() {
-    const proj = this.route.snapshot.queryParams['proj'];
-    if (proj) {
-      this.project = proj;
-    }
+  override getBackendData(): Observable<string> {
     const paramMap = this.route.snapshot.paramMap;
-
     if (paramMap.get('projectId') !== null) {
-      // user accesses project text by entering a url manually
-      this.getDataService.getProjectTextData(paramMap).subscribe((data) => {
-        this.handleTextToHTMLConversion(data, true);
-      });
-    } else {
-      this.getDataService.getDetailData(
-        this.project,
-        this.route.snapshot.queryParams['lang'],
-        this.route.snapshot.queryParams['isid'],
-        {ref: this.route.snapshot.queryParams['iref']}
-      ).subscribe((data) => {
-        this.handleTextToHTMLConversionP4(data, true);
-      });
+      // User accesses project text by entering a url manually.
+      // This won't work...
+      return this.getDataService.getProjectTextData(paramMap);
     }
-    this.chosenTermText = this.route.snapshot.paramMap.get('word');
-    this.isMobile = window.innerWidth < 991 ? true : false;
-  }
-
-  private handleTextToHTMLConversion(text: string, isTermData = false) {
-    const parser = new DOMParser();
-    const htmlData = parser.parseFromString(text, 'text/html');
-    const htmlDataToBeReduced = parser.parseFromString(text, 'text/html');
-    const metadataPanelInput = htmlData.getElementById('p3left');
-    const middlePanelInput = htmlData.getElementById('p3right');
-    const textPanelInput = htmlDataToBeReduced.getElementById('p3right');
-
-    middlePanelInput.querySelectorAll('td').forEach((node) => {
-      if (node.className === 't1 xtr') {
-        if (typeof node.remove === 'function') {
-          node.remove();
-        } else {
-          node.outerHTML = '';
-        }
-      }
-    });
-    textPanelInput.querySelectorAll('td').forEach((node) => {
-      if (node.className !== 't1 xtr') {
-        if (typeof node.remove === 'function') {
-          node.remove();
-        } else {
-          node.outerHTML = '';
-        }
-      }
-    });
-
-    const controlsInput = htmlData.getElementById('p4PageNav');
-    this.totalLines = parseInt(
-      controlsInput.getAttribute('data-imax'),
-      10
-    );
-    const pageCount = parseInt(
-      controlsInput.getAttribute('data-pmax'),
-      10
-    );
-    this.totalPages = Array(pageCount - this.paginationSliceStart)
-      .fill(1)
-      .map((e, i) => i + 1);
-    this.shouldShowPaginationArrows = this.totalPages.length > 6;
-    this.paginatedPages = this.totalPages.slice(
-      this.paginationSliceStart,
-      this.paginationSliceEnd
-    );
-
-    this.metadataPanel = "<i class='fas fa-spinner'></i>";
-    this.metadataPanel = this.sanitizer.bypassSecurityTrustHtml(
-      metadataPanelInput.innerHTML
-    );
-
-    this.middlePanel = "<i class='fas fa-spinner'></i>";
-    this.middlePanel = this.sanitizer.bypassSecurityTrustHtml(
-      middlePanelInput.innerHTML
-    );
-
-    this.textPanel = "<i class='fas fa-spinner'></i>";
-    this.textPanel = this.sanitizer.bypassSecurityTrustHtml(
-      textPanelInput.innerHTML
+    return this.getDataService.getDetailData(
+      this.project,
+      this.route.snapshot.queryParams['lang'],
+      this.route.snapshot.queryParams['isid'],
+      {ref: this.route.snapshot.queryParams['iref']}
     );
   }
 
-  private handleTextToHTMLConversionP4(text: string, isTermData = false) {
-    const parser = new DOMParser();
-    const htmlData = parser.parseFromString(text, 'text/html');
-    const pager: HTMLElement = htmlData.getElementById('p4Pager');
-    this.item = pager.getAttribute('data-item');
-    const outlineInput = htmlData.getElementById('p4MenuOutline');
+  override setMetadataPanel(htmlData: Document) {
     const metadataInput = htmlData.getElementById('p4XtfMeta');
-
-    const controlsInput = htmlData.getElementById('p4PageNav');
-
-    this.metadataPanel = "<i class='fas fa-spinner'></i>";
+    const outlineInput = htmlData.getElementById('p4MenuOutline');
     let m = document.createElement('div');
     m.append(metadataInput, outlineInput);
     this.metadataPanel = this.sanitizer.bypassSecurityTrustHtml(m.innerHTML);
+  }
 
+  override setMiddlePanel(htmlData : Document) {
+    const pager: HTMLElement = htmlData.getElementById('p4Pager');
+    this.item = pager.getAttribute('data-item');
     this.handleTextToHTMLConversionText(htmlData, 'p4XtfData');
   }
 
   private handleTextToHTMLConversionText(htmlData: Document, middleId: string) {
     let middlePanelInput = htmlData.getElementById(middleId);
-    this.middlePanel = "<i class='fas fa-spinner'></i>";
-    this.textPanel = "<i class='fas fa-spinner'></i>";
     const textPanelInput = splitOutTranslations(middlePanelInput);
     this.middlePanel = this.sanitizer.bypassSecurityTrustHtml(
       middlePanelInput.innerHTML
     );
     this.textPanel = this.sanitizer.bypassSecurityTrustHtml(
       textPanelInput.innerHTML
-    );
-  }
-
-  private handleTextToHTMLConversionOnPageChange(text: string) {
-    const parser = new DOMParser();
-    const htmlData = parser.parseFromString(text, 'text/html');
-    const middlePanelInput = htmlData.querySelector('body');
-
-    this.middlePanel = this.sanitizer.bypassSecurityTrustHtml(
-      middlePanelInput.innerHTML
     );
   }
 
@@ -289,63 +175,6 @@ export class DetailsTextsComponent implements OnInit {
       return;
     }
     window.open(clickedLink.getAttribute('href'));
-  }
-
-  public togglePanel(e, panelType) {
-    if (panelType === PANEL_TYPE.METADATA) {
-      this.isMetadataPanelActive = !this.isMetadataPanelActive;
-    }
-    if (panelType === PANEL_TYPE.TEXT) {
-      this.isTextPanelActive = !this.isTextPanelActive;
-    }
-  }
-
-  private handlePaginationBoundary(currentPage: number) {
-    const indexOfCurrentPage = this.paginatedPages.indexOf(currentPage);
-    if (
-      indexOfCurrentPage === 6 &&
-      currentPage < this.totalPages[this.totalPages.length - 1]
-    ) {
-      this.paginatedPages = this.totalPages.slice(
-        (this.paginationSliceStart += 1),
-        (this.paginationSliceEnd += 1)
-      );
-    }
-    if (indexOfCurrentPage === 0 && currentPage > 1) {
-      this.paginatedPages = this.totalPages.slice(
-        (this.paginationSliceStart -= 1),
-        (this.paginationSliceEnd -= 1)
-      );
-    }
-  }
-
-  public handlePageChange(e, direction?: string) {
-    const oldPage = this.currentPage;
-    if (!direction) {
-      this.currentPage = parseInt(e.target.innerHTML, 10);
-    }
-    if (direction === DIRECTION.BACK && this.currentPage >= 1) {
-      this.currentPage = this.currentPage - 1 || 1;
-    }
-    if (
-      direction === DIRECTION.FORWARD &&
-      this.totalPages.length !== this.currentPage
-    ) {
-      this.currentPage = this.currentPage + 1;
-    }
-    if (oldPage !== this.currentPage) {
-      this.handlePaginationBoundary(this.currentPage);
-      this.getDataService.getDetailData(
-        this.project,
-        this.route.snapshot.queryParams['ga_lang'],
-        this.route.snapshot.queryParams['ga_isid'],
-        { page: this.currentPage,
-          ref: this.route.snapshot.queryParams['iref']
-        }
-      ).subscribe((data) => {
-        this.handleTextToHTMLConversionOnPageChange(data);
-      });
-    }
   }
 
   public handleTranslationClick(e) {
