+++ src/app/pages/details/details.component.ts
@@ -1,10 +1,6 @@
-import { Component, OnInit, ViewEncapsulation } from '@angular/core';
-import { GetDataService } from '../../services/get-data/get-data.service';
-import { HandleBreadcrumbsService } from 'src/app/services/handle-breadcrumbs/handle-breadcrumbs.service';
-import { DomSanitizer } from '@angular/platform-browser';
-import { ActivatedRoute, ParamMap, Params, Router } from '@angular/router';
+import { Component, ViewEncapsulation } from '@angular/core';
 import { composedPath } from '../../../utils/utils';
-import { DIRECTION, PANEL_TYPE } from '../../../utils/consts';
+import { ThreePanel } from 'src/app/components/three-panel.component';
 
 @Component({
   selector: 'app-details',
@@ -12,90 +8,21 @@ import { DIRECTION, PANEL_TYPE } from '../../../utils/consts';
   styleUrls: ['./details.component.scss'],
   encapsulation: ViewEncapsulation.None
 })
-export class DetailsComponent implements OnInit {
-  public metadataPanel: any = "<i class='fas fa-spinner'></i>";
-  public chosenTermText: string = "";
-  public currentPage = 1;
-  public textPanel: any;
-  public isMetadataPanelActive = window.innerWidth > 991 ? true : false;
-  public totalLines: number;
-  public totalPages: any;
-  public paginatedPages: any;
-  public middlePanel: any = "<i class='fas fa-spinner'></i>";
-  public isTextPanelActive = window.innerWidth > 991 ? true : false;
-  public shouldShowPaginationArrows: boolean;
-  private paginationSliceStart: number = 0;
-  private paginationSliceEnd: number = 7;
-  private project: string = 'neo';
-  private zoom: number | null = null;
-
-  constructor(
-    private getDataService: GetDataService,
-    private breadcrumbsService: HandleBreadcrumbsService,
-    private sanitizer: DomSanitizer,
-    private router: Router,
-    private route: ActivatedRoute
-  ) {
-    this.breadcrumbsService.setBreadcrumbs(this.router);
-  }
-
-  ngOnInit() {
-    const proj = this.route.snapshot.queryParams['proj'];
-    if (proj) {
-      this.project = proj;
-    }
-    this.getDataService.getDetailData(
-      this.project,
-      this.route.snapshot.queryParams['lang'],
-      this.route.snapshot.queryParams['isid'],
-      {}
-    ).subscribe((data) => {
-      this.handleTextToHTMLConversion(data);
-    });
-    this.chosenTermText = this.route.snapshot.paramMap.get('word');
-  }
-
-  private setMiddlePanel(htmlData : Document) {
+export class DetailsComponent extends ThreePanel {
+  override setMiddlePanel(htmlData : Document) {
     const middlePanelInput = htmlData.getElementById('p4Content');
 
-    const controlsInput = htmlData.getElementById('p4PageNav');
-
-    this.totalLines = parseInt(
-      controlsInput.getAttribute('data-imax'),
-      10
-    );
-    const pageCount = parseInt(
-      controlsInput.getAttribute('data-pmax'),
-      10
-    );
-    this.totalPages = Array(pageCount - this.paginationSliceStart)
-      .fill(1)
-      .map((e, i) => i + 1);
-    this.shouldShowPaginationArrows = this.totalPages.length > 6;
-    this.paginatedPages = this.totalPages.slice(
-      this.paginationSliceStart,
-      this.paginationSliceEnd
-    );
     this.middlePanel = this.sanitizer.bypassSecurityTrustHtml(
       middlePanelInput.innerHTML
     );
+    this.setTextPanel(htmlData);
   }
 
-  private handleTextToHTMLConversion(text: string) {
-    const parser = new DOMParser();
-    const htmlData = parser.parseFromString(text, 'text/html');
+  override setMetadataPanel(htmlData: Document) {
     const metadataPanelInput = htmlData.getElementById('p4MenuOutline');
-
     this.metadataPanel = this.sanitizer.bypassSecurityTrustHtml(
       metadataPanelInput.innerHTML
     );
-    this.setMiddlePanel(htmlData);
-  }
-
-  private handleTextToHTMLConversionOnPageChange(text) {
-    const parser = new DOMParser();
-    const htmlData = parser.parseFromString(text, 'text/html');
-    this.setMiddlePanel(htmlData);
   }
 
   public handleDetailsClick(e) {
@@ -130,79 +57,10 @@ export class DetailsComponent implements OnInit {
     }
   }
 
-  private updateForZoom() {
-    this.getDataService.getDetailData(
-      this.project,
-      this.route.snapshot.queryParams['lang'],
-      this.route.snapshot.queryParams['isid'], {
-        page: this.currentPage,
-        zoom: this.zoom
-      }
-    ).subscribe(data => this.handleTextToHTMLConversionOnPageChange(data));
-  }
-
   public handleMetadataClick(e) {
     e.preventDefault();
     if (e.target.tagName === "A" && e.target.hasAttribute('data-zoom')) {
-      this.zoom = parseInt(e.target.getAttribute('data-zoom'), 10);
-      this.updateForZoom();
-    }
-  }
-
-  public handleResetZoom(e) {
-    this.zoom = null;
-    this.updateForZoom();
-  }
-
-  public isZoomed() {
-    return this.zoom != null;
-  }
-
-  public togglePanel(e, panelType) {
-    if (panelType === PANEL_TYPE.METADATA) {
-      this.isMetadataPanelActive = !this.isMetadataPanelActive;
-    }
-    if (panelType === PANEL_TYPE.TEXT) {
-      this.isTextPanelActive = !this.isTextPanelActive;
-    }
-  }
-
-  private handlePaginationBoundary(currentPage: number) {
-    const indexOfCurrentPage = this.paginatedPages.indexOf(currentPage);
-    if (
-      indexOfCurrentPage === 6 &&
-      currentPage < this.totalPages[this.totalPages.length - 1]
-    ) {
-      this.paginatedPages = this.totalPages.slice(
-        (this.paginationSliceStart += 1),
-        (this.paginationSliceEnd += 1)
-      );
-    }
-    if (indexOfCurrentPage === 0 && currentPage > 1) {
-      this.paginatedPages = this.totalPages.slice(
-        (this.paginationSliceStart -= 1),
-        (this.paginationSliceEnd -= 1)
-      );
-    }
-  }
-
-  public handlePageChange(e, direction?: string) {
-    const oldPage = this.currentPage;
-    if (!direction) {
-      this.currentPage = parseInt(e.target.innerHTML, 10);
-    }
-    if (direction === DIRECTION.BACK && this.currentPage >= 1) {
-      this.currentPage = this.currentPage - 1 || 1;
-    }
-    if (
-      direction === DIRECTION.FORWARD &&
-      this.totalPages.length !== this.currentPage
-    ) {
-      this.currentPage = this.currentPage + 1;
-    }
-    if (oldPage !== this.currentPage) {
-      this.handlePaginationBoundary(this.currentPage);
-      this.updateForZoom();
+      this.setZoom(parseInt(e.target.getAttribute('data-zoom'), 10));
     }
   }
 }
